+++ actiontext/lib/action_text/attachables/content_attachment.rb
@@ -21,6 +21,10 @@ module ActionText
         content_instance.fragment.source
       end
 
+      def attachable_markdown_representation(caption)
+        content_instance.fragment.to_markdown
+      end
+
       def to_html
         @to_html ||= content_instance.render(content_instance)
       end